Update README.md
In order to protect ourselves from the COVID-19 Pandemic, almost every one of us tend to wear a face mask. It becomes increasingly necessary to check if the people in the crowd wear face masks, In my turn, I tried to add a voice alert showing that access is denied for people who do not wear a mask. This kind of system, has applications in most public gatherings such as Malls, Theatres, Parks . The development of an AI solution to detect if the person is wearing a face mask and allow their entry would be of great help to the society. In this,  Real Time Face Mask Detection And Alert system is built using the Deep Learning technique called as Convolutional Neural Networks (CNN). This CNN Model is built using the TensorFlow framework and the OpenCV library which is highly used for real-time applications. This model can also be used to develop a full-fledged software to scan every person before they can enter the public gathering. Using this model, an accuracy of over 95% is obtained. This can also be used further to achieve even higher levels of accuracy.